Staging: fixed DNS validation
