--- EmailValidator/Exception/NoDNSRecord.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace Egulias\EmailValidator\Exception;
-
-class NoDNSRecord extends InvalidEmail
-{
-    const CODE = 5;
-    const REASON = 'No MX or A DSN record was found for this email';
-}
