+++ EmailValidator/Parser/DomainPart.php
@@ -3,16 +3,13 @@
 namespace Egulias\EmailValidator\Parser;
 
 use Egulias\EmailValidator\EmailLexer;
-use Egulias\EmailValidator\Exception\CharNotAllowed;
 use Egulias\EmailValidator\Exception\CommaInDomain;
 use Egulias\EmailValidator\Exception\ConsecutiveAt;
 use Egulias\EmailValidator\Exception\CRNoLF;
 use Egulias\EmailValidator\Exception\ExpectingATEXT;
-use Egulias\EmailValidator\Exception\ExpectingDomainLiteralClose;
 use Egulias\EmailValidator\Exception\ExpectingDTEXT;
 use Egulias\EmailValidator\Result\InvalidEmail;
 use Egulias\EmailValidator\Result\Reason\CharNotAllowed as ReasonCharNotAllowed;
-use Egulias\EmailValidator\Result\Reason\CRLFAtTheEnd as ReasonCRLFAtTheEnd;
 use Egulias\EmailValidator\Result\Reason\DomainHyphened as ReasonDomainHyphened;
 use Egulias\EmailValidator\Result\Reason\DotAtEnd as ReasonDotAtEnd;
 use Egulias\EmailValidator\Result\Reason\DotAtStart;
@@ -26,7 +23,6 @@ use Egulias\EmailValidator\Warning\DomainLiteral;
 use Egulias\EmailValidator\Warning\DomainTooLong;
 use Egulias\EmailValidator\Warning\IPV6BadChar;
 use Egulias\EmailValidator\Warning\IPV6ColonEnd;
-use Egulias\EmailValidator\Warning\IPV6ColonStart;
 use Egulias\EmailValidator\Warning\IPV6Deprecated;
 use Egulias\EmailValidator\Warning\IPV6DoubleColon;
 use Egulias\EmailValidator\Warning\IPV6GroupCount;
@@ -34,6 +30,8 @@ use Egulias\EmailValidator\Warning\IPV6MaxGroups;
 use Egulias\EmailValidator\Warning\LabelTooLong;
 use Egulias\EmailValidator\Warning\ObsoleteDTEXT;
 use Egulias\EmailValidator\Warning\TLD;
+use Egulias\EmailValidator\Parser\DomainLiteral as DomainLiteralParser;
+use Egulias\EmailValidator\Result\Reason\ExpectingDomainLiteralClose;
 
 class DomainPart extends Parser
 {
@@ -124,53 +122,6 @@ class DomainPart extends Parser
         return $this->domainPart;
     }
 
-    /**
-     * @param string $addressLiteral
-     * @param int $maxGroups
-     */
-    public function checkIPV6Tag($addressLiteral, $maxGroups = 8)
-    {
-        $prev = $this->lexer->getPrevious();
-        if ($prev['type'] === EmailLexer::S_COLON) {
-            $this->warnings[IPV6ColonEnd::CODE] = new IPV6ColonEnd();
-        }
-
-        $IPv6       = substr($addressLiteral, 5);
-        //Daniel Marschall's new IPv6 testing strategy
-        $matchesIP  = explode(':', $IPv6);
-        $groupCount = count($matchesIP);
-        $colons     = strpos($IPv6, '::');
-
-        if (count(preg_grep('/^[0-9A-Fa-f]{0,4}$/', $matchesIP, PREG_GREP_INVERT)) !== 0) {
-            $this->warnings[IPV6BadChar::CODE] = new IPV6BadChar();
-        }
-
-        if ($colons === false) {
-            // We need exactly the right number of groups
-            if ($groupCount !== $maxGroups) {
-                $this->warnings[IPV6GroupCount::CODE] = new IPV6GroupCount();
-            }
-            return;
-        }
-
-        if ($colons !== strrpos($IPv6, '::')) {
-            $this->warnings[IPV6DoubleColon::CODE] = new IPV6DoubleColon();
-            return;
-        }
-
-        if ($colons === 0 || $colons === (strlen($IPv6) - 2)) {
-            // RFC 4291 allows :: at the start or end of an address
-            //with 7 other groups in addition
-            ++$maxGroups;
-        }
-
-        if ($groupCount > $maxGroups) {
-            $this->warnings[IPV6MaxGroups::CODE] = new IPV6MaxGroups();
-        } elseif ($groupCount === $maxGroups) {
-            $this->warnings[IPV6Deprecated::CODE] = new IPV6Deprecated();
-        }
-    }
-
     protected function parseComments()
     {
         $commentParser = new Comment($this->lexer, new DomainComment());
@@ -210,8 +161,12 @@ class DomainPart extends Parser
                 return $result;
             }
 
-            if ($this->hasBrackets()) {
-                $this->parseDomainLiteral();
+            if ($this->lexer->token['type'] !== EmailLexer::S_OPENBRACKET) {
+                $literalResult = $this->parseDomainLiteral();
+                //Invalid literal parsing
+                if($literalResult->isInvalid()) {
+                    return $literalResult;
+                }
             }
 
             $this->checkLabelLength($prev);
@@ -240,118 +195,21 @@ class DomainPart extends Parser
     /**
      * @return string|false
      */
-    protected function parseDomainLiteral()
+    protected function parseDomainLiteral() : Result
     {
-        if ($this->lexer->isNextToken(EmailLexer::S_COLON)) {
-            $this->warnings[IPV6ColonStart::CODE] = new IPV6ColonStart();
-        }
-        if ($this->lexer->isNextToken(EmailLexer::S_IPV6TAG)) {
-            $lexer = clone $this->lexer;
-            $lexer->moveNext();
-            if ($lexer->isNextToken(EmailLexer::S_DOUBLECOLON)) {
-                $this->warnings[IPV6ColonStart::CODE] = new IPV6ColonStart();
-            }
+
+        try {
+            $this->lexer->find(EmailLexer::S_CLOSEBRACKET);
+        } catch (\RuntimeException $e) {
+            return new InvalidEmail(new ExpectingDomainLiteralClose(), $this->lexer->token['value']);
         }
 
-        return $this->doParseDomainLiteral();
-    }
+        $domainLiteralParser = new DomainLiteralParser($this->lexer);
+        $result = $domainLiteralParser->parse('remove');
+        $this->warnings = array_merge($this->warnings, $domainLiteralParser->getWarnings());
+        return $result;
 
-    /**
-     * @return string|false
-     */
-    protected function doParseDomainLiteral()
-    {
-        $IPv6TAG = false;
-        $addressLiteral = '';
-        do {
-            if ($this->lexer->token['type'] === EmailLexer::C_NUL) {
-                throw new ExpectingDTEXT();
-            }
-
-            if ($this->lexer->token['type'] === EmailLexer::INVALID ||
-                $this->lexer->token['type'] === EmailLexer::C_DEL   ||
-                $this->lexer->token['type'] === EmailLexer::S_LF
-            ) {
-                $this->warnings[ObsoleteDTEXT::CODE] = new ObsoleteDTEXT();
-            }
-
-            if ($this->lexer->isNextTokenAny(array(EmailLexer::S_OPENQBRACKET, EmailLexer::S_OPENBRACKET))) {
-                throw new ExpectingDTEXT();
-            }
-
-            if ($this->lexer->isNextTokenAny(
-                array(EmailLexer::S_HTAB, EmailLexer::S_SP, $this->lexer->token['type'] === EmailLexer::CRLF)
-            )) {
-                $this->warnings[CFWSWithFWS::CODE] = new CFWSWithFWS();
-                $this->parseFWS();
-            }
-
-            if ($this->lexer->isNextToken(EmailLexer::S_CR)) {
-                throw new CRNoLF();
-            }
-
-            if ($this->lexer->token['type'] === EmailLexer::S_BACKSLASH) {
-                $this->warnings[ObsoleteDTEXT::CODE] = new ObsoleteDTEXT();
-                $addressLiteral .= $this->lexer->token['value'];
-                $this->lexer->moveNext();
-                $this->validateQuotedPair();
-            }
-            if ($this->lexer->token['type'] === EmailLexer::S_IPV6TAG) {
-                $IPv6TAG = true;
-            }
-            if ($this->lexer->token['type'] === EmailLexer::S_CLOSEQBRACKET) {
-                break;
-            }
-
-            $addressLiteral .= $this->lexer->token['value'];
-
-        } while ($this->lexer->moveNext());
-
-        $addressLiteral = str_replace('[', '', $addressLiteral);
-        $addressLiteral = $this->checkIPV4Tag($addressLiteral);
-
-        if (false === $addressLiteral) {
-            return $addressLiteral;
-        }
-
-        if (!$IPv6TAG) {
-            $this->warnings[DomainLiteral::CODE] = new DomainLiteral();
-            return $addressLiteral;
-        }
-
-        $this->warnings[AddressLiteral::CODE] = new AddressLiteral();
-
-        $this->checkIPV6Tag($addressLiteral);
-
-        return $addressLiteral;
-    }
-
-    /**
-     * @param string $addressLiteral
-     *
-     * @return string|false
-     */
-    protected function checkIPV4Tag($addressLiteral)
-    {
-        $matchesIP  = array();
-
-        // Extract IPv4 part from the end of the address-literal (if there is one)
-        if (preg_match(
-            '/\\b(?:(?:25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)\\.){3}(?:25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)$/',
-            $addressLiteral,
-            $matchesIP
-        ) > 0
-        ) {
-            $index = strrpos($addressLiteral, $matchesIP[0]);
-            if ($index === 0) {
-                $this->warnings[AddressLiteral::CODE] = new AddressLiteral();
-                return false;
-            }
-            // Convert IPv4 part to IPv6 format for further testing
-            $addressLiteral = substr($addressLiteral, 0, (int) $index) . '0:0';
-        }
-
-        return $addressLiteral;
+        //return $this->doParseDomainLiteral();
     }
 
     protected function checkDomainPartExceptions(array $prev)
@@ -389,24 +247,6 @@ class DomainPart extends Parser
         }
 
         return new ValidEmail();
-    }
-
-    /**
-     * @return bool
-     */
-    protected function hasBrackets()
-    {
-        if ($this->lexer->token['type'] !== EmailLexer::S_OPENBRACKET) {
-            return false;
-        }
-
-        try {
-            $this->lexer->find(EmailLexer::S_CLOSEBRACKET);
-        } catch (\RuntimeException $e) {
-            throw new ExpectingDomainLiteralClose();
-        }
-
-        return true;
     }
 
     protected function checkLabelLength(array $prev)
