+++ EmailValidator/Result/Reason/ExpectingDTEXT.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace Egulias\EmailValidator\Result\Reason;
+
+class ExpectingDTEXT implements Reason
+{
+    public function code() : int
+    {
+        return 129;
+    }
+
+    public function description() : string
+    {
+        return 'Expecting DTEXT';
+    }
+}
