+++ EmailValidator/Result/Reason/NoDNSRecord.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace Egulias\EmailValidator\Result\Reason;
+
+class NoDNSRecord implements Reason 
+{
+    public function code() : int
+    {
+        return 5;
+    }
+
+    public function description() : string
+    {
+        return 'No MX or A DSN record was found for this email';
+    }
+}
