+++ EmailValidator/Result/Reason/RFCWarnings.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace Egulias\EmailValidator\Result\Reason;
+
+class RFCWarnings implements Reason
+{
+    public function code() : int
+    {
+        return 997;
+    }
+
+    public function description() : string
+    {
+        return 'Warnings found after validating';
+    }
+}
