+++ Tests/EmailValidator/Validation/NoRFCWarningsValidationTest.php
@@ -3,7 +3,6 @@
 namespace Egulias\Tests\EmailValidator\Validation;
 
 use Egulias\EmailValidator\EmailLexer;
-use Egulias\EmailValidator\Exception\NoDomainPart;
 use Egulias\EmailValidator\Result\Reason\NoDomainPart as ReasonNoDomainPart;
 use Egulias\EmailValidator\Validation\Error\RFCWarnings;
 use Egulias\EmailValidator\Validation\NoRFCWarningsValidation;
